openai_patching/requirements.txt
openai
pymupdf
python-dotenv
paddleocr
pillow
numpy
PyMuPDF
paddlepaddle -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
sentence-transformers
torch